requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-4-jet==1.0.8
django-ckeditor==6.2.0
django-jet==1.0.8
django-js-asset==2.0.0
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
google-api-python-client==1.4.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
oauth2client==4.1.3
oauthlib==3.2.0
phonenumbers==8.12.43
Pillow==9.0.1
# 2.7.7
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
